Made small error corrections, mostly in comments
These changes do not change the behaviour of the code in any way --
except in one case, where it solves a very rare problem (param.py).
Some comments and doc-strings are corrected, an unused variable is
removed from vortex.py and a confusing line in shear_instability.py was
simplified and explained by adding a comment.