Replace geometry "square" by "closed"
Before this commit, the names "square" and "closed" were both used for
the same domain geometry.  However, square was used in many situations,
where the domain is not actually squared, but rectangular.  To avoid this
potential confusion, all remaining occurrences of "square" are replaced
by "closed".  Now, the geometry "square" cannot be used anymore.