Add new feature to fluid2d: EMS
EMS stands for Experiment Management System and is a new way to handle
big sets of experiments.  It is an optional feature that does not
influence the way fluid2d works unless activated by the user.  An
example of its usage is shown in the new experiment `Waves with EMS`.