EMShell: auto-set optimal width for comments
The last column in the table view of the EMShell, which displays the
comment of an experiment, can now automatically adopt its width to the
width available in the terminal window.  Also the full comment man be
displayed every time instead.

The option to start the EMShell without fluid2d activated is removed.
It did not fit well into the general behaviour of the fluid2d-system.

New function is added to the EMShell to display free disk space.