EMShell: print datetime in an easy-to-read format
A new setting is added to the EMShell which, when activated, prints the
value of the datetime column in a format which is easily readable by
humans, for example by saying "15 min ago" instead of the exact date and
time.
The ems was slightly modified to prevent unexpected behaviour in the
rare case that datetime.now() returns a time with 0 microseconds.