Clean up the implementation of forcings
The parameter forcing_module has now the default value 'embedded'.
Therefore it is not needed to mention the forcing_module at all if it is
embedded in the python file.  It is considered a left-over for backwards
compatibility, which can be useful in cases where the definition of the
forcing is very big, as in the Rayleigh--Benard experiment.
The other two experiments with forcing (TwoDimTurb and GyreCirculation)
now feature the new, clean, slim way to implement a forcing.  They can
be considered a good example for the usage of forcings.

The warning message for embedded forcings is corrected.

The file core/forcing.py was left over, not needed anymore and is
removed now.