EMShell: Add new functionality and unify behaviour
With this commit it is possible to remove tables from the database and
to change to comment of entries in the database.  The behaviour and the
help texts of several commands are made uniform.  Instead of an ID, it
is now possible to say "last" or "-1".

A bug in the ems is fixed which caused a crash when a table was empty.