Fix output when fluid2d runs on multiple cores
When fluid2d runs on multiple cores, the name of the his file displayed
at the end was wrong.  The filename that was printed after the
simulation finished was the name of one of the his-files which was
joined into one big his-file and then removed.  With this commit,
fluid2d prints the correct name of the his-file after the run even when
using multiple cores.  The same is true for the flux file.

Creating an mp4-file while fluid2d runs on multiple cores was possible
so far, however, it resulted in the creation of just one file which
shows only some part of the domain, not fully.  Therefore it was only
limited usefully.  Furthermore it led to a crash, since several
fluid2d-processes called "finalize" on the same file.  Also creating an
mp4-file, which is not really helpful wastes resouces.  Therefore, this
commit disables the possbility to generate mp4s when fluid2d runs on
multiple cores.  A warning is printed instead.

The member variables of Output hisfile_joined and flxfile_joined were
unused and are now removed.