EMS: Implement support for multi values and cores
The EMS is almost fully reworked in this commit.  It now contains the
long desired feature of specifying multiple values in the experiment
file and automatically running one experiment for every possible
combination of these values.  Once again, this is demonstrated in the
Waves experiment.  Furthermore, the EMS supports now natively
multi-threading, which required some manual hacks before.  However, this
required to add the EMS to the Param class, which, at the end, made the
whole implementation a lot cleaner and easier for the user.
In consequence, this commit modified, for the first time, files of the
standard distribution of fluid2d.  The influence was kept as tiny as
possbile.  As before, the EMS does not make a difference for the user
unless he activates it, which brings now a whole new experience!