Add new feature ‘custom diagnostics’ to Fluid2D
Custom diagnostics allow to store any information that can be extracted
from the model state in the diagnostics file.  With this, we can provide
a high frequency output of a diagnostics we are interested in without
creating a huge history file.  This may make the diagnostics file more
useful for a wider range of Fluid2D users.

An example is provided where custom diagnostics are used to measure the
drift of a vortex that he experiences due to the beta-effect on a
rotating planet.

Note: the experiment only works correctly with the fixed QG model
(see pull request "Fix QG model")