EMS: Improve code and add new functionality to modify columns
This commit brings a lot of code improvements, most of them in the
DB-connection class, which was simplified a lot.

Furthermore, new functionality was implemented to add or remove columns
from a table.

The requirement to have numpy installed was removed from the
EMShell (apart from the EMShellExtensions, which are considered
optional) and matplotlib was made an optional requirement.  A lot (but
not all of the functionality of EMS) can be used without matplotlib, so
it is now possible to run EMShell on a computer without matplotlib.

When matplotlib is used, a different backend is now chosen by default,
which allows the user to modify the plot manually.  This is a big
advantage when the user wants to create nice figures.

Also a new colormap was introduced that can be chosen by the user.

Finally, comments were added and improved.